+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
   - Process It blocks in memory [GH-123]
   - Fixed -ExecutionPolicy in pester.bat [GH-130]
+  - Add support for mocking internal module functions, aliases, exe and filters. [GH-126]
 
 ## 2.0.4 (March 9, 2014)